--- src/components/TestForm1/index.scss
@@ -1,12 +1,0 @@
-body{
-    background-color: #e9e9e9;
-}
-
-.test-form-field{
-    background-color:white;
-    border-radius: 5px;
-    padding: 35px;
-    margin: 25px;
-}
-
-
