--- src/components/TestForm2/index.scss
@@ -1,19 +1,0 @@
-@import "../Styles/variables";
-@import "../Styles/mixins";
-@import "../Styles/extends";
-
-body{
-    background-color: #e9e9e9;
-}
-
-.registration-field {
-    background-color: white;
-    border-radius: 5px;
-    padding: 35px;
-    margin: 25px;
-}
-
-
-
-
-
