+++ src/lib/ml-form-builder/lib/Styles/_extends.scss
@@ -1,0 +1,30 @@
+%helper-error-text-size-opacity{
+    font-size: $helperErrorTextSize;
+    opacity: $primaryOpacity;
+}
+%label-text-size-opacity{
+    font-size: $labelTextSize;
+    opacity: $primaryOpacity;
+}
+
+%btn{
+  font-size: $labelTextSize;
+  background-color: $labelTextColor;
+  border-radius: 4px;
+  color:white ;
+  padding: 14px 30px;
+  text-align: center;
+  cursor: pointer;
+  margin-top: 5px;
+  border: none;
+}
+
+%input-select-box{
+    box-sizing: border-box;
+    border: none;
+    outline: none;
+    width: 100%;
+    height: 30px;
+    padding: 0px;
+    border-bottom: 1px solid $labelTextColor;
+}
