+++ src/lib/ml-form-builder/lib/Switch/index.tsx
@@ -1,0 +1,49 @@
+import React from "react";
+import { FieldItemProps, FieldProps } from "../Types";
+import "./index.scss";
+import { get } from "lodash";
+import { FormikProps } from "formik";
+import clsx from "clsx";
+import HelperText from "../HelperText";
+
+export interface SwitchFieldProps extends FieldItemProps {
+  label?: string;
+}
+
+interface SwitchProps extends FieldProps {
+  fieldProps?: SwitchFieldProps & FieldItemProps;
+}
+
+const Switch: React.FC<SwitchProps> = (props) => {
+  const {
+    formikProps = {} as FormikProps<unknown>,
+    fieldProps = {} as SwitchFieldProps,
+  } = props;
+  const { label, name = "", helperText, classNames, nativeProps, disabled} = fieldProps;
+
+  const fieldValue = get(formikProps, `values.${name}`);
+  const handleOnChange = () => {
+    formikProps.setFieldValue(`${name}`, !fieldValue);
+  };
+  return (
+    <div className={clsx("switch-field", classNames)}>
+      {label && <span className="switch-label switchlabel">{label}</span>}
+      <label className="switch-container">
+        <input
+          className="slider"
+          type="checkbox"
+          checked={!!fieldValue}
+          value={fieldValue}
+          onChange={handleOnChange}
+          disabled={disabled}
+          {...nativeProps}
+        />
+        <span className="slider round"></span>
+      </label>
+
+      <HelperText fieldProps={fieldProps} formikProps={formikProps} />
+    </div>
+  );
+};
+
+export default Switch;
